requirements.txt
fastapi
uvicorn
motor
bcrypt
python-dotenv
python-jose[cryptography]
pydantic
